requirements.txt
# ThermalOS Digital Twin MVP – Python-Abhängigkeiten
# Bereinigt: pydantic, fastapi, uvicorn und scipy entfernt (nicht verwendet)

streamlit>=1.30.0
plotly>=5.18.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
paho-mqtt>=1.6.1
influxdb-client>=1.38.0
pyyaml>=6.0
